Move global shortcut hooks to App
